1834. Single threaded cpu

+++ coding_challenges/leetcode/heaps/1834_single_threaded_cpu.cc
@@ -11,7 +11,7 @@ std::vector<int> getOrder(std::vector<std::vector<int>>& tasks) {
 
     std::vector<int> res;
     res.reserve(N);
-    int running_time = 0, i = 0;
+    long running_time = 0, i = 0;
 
     for (int i : std::views::iota(0, N)) tasks[i].push_back(i);
 
@@ -19,7 +19,7 @@ std::vector<int> getOrder(std::vector<std::vector<int>>& tasks) {
 
     while (i < N || pq.size() > 0) {
         if (pq.empty()) {
-            running_time = std::max(running_time, tasks[i][0]);
+            running_time = std::max(running_time, (long)tasks[i][0]);
         }
 
         while (i < N || running_time >= tasks[i][0]) {
